unique_ptr<WriterLoop> -> WriterLoop
Summary:
There's no need for it to be allocated separately and we access it for every request (except batched once). This way we avoid loading it from memery on access (which usually happens righ after checkin `writeScheduled_`.
I can see `sendCommon` goin down in perf record from 1.2% to 1%, also `pushMessages` is slightly better.

Reviewed By: jmswen

Differential Revision: D5782297

fbshipit-source-id: 615f18523a15c83b337a73be0a7ba8a5e6d5c628