Smarter notifications, batching & scheduling
Summary: Over the week I was thinking how to avoid double-scheduling `WriterLoop` (i.e. end of next loop). After jmswen fix to the messagequeue I realized that these things are related and we can use knowledge of them to improve each other, i.e. we ever need to set `State::EMPTY` if only we're going to block, similarly we don't need to flush every 2nd loop, but instead when we know that there's nothing more to do.

Reviewed By: jmswen

Differential Revision: D5802302

fbshipit-source-id: a4a7559843e5cb397d8a7b1e580f5843bb650bb7