Strengthen usage of LeaseTokenMap in FailoverRoute
Summary:
The current way that LeaseTokenMap is being used in FailoverRoute, allows for an
out-of-range bug to be triggered if all of the following happens:
1) There are multiple bad hosts
2) We failover LeaseGet deeply (likely to the last possible destinations).
3) We reconfigure and in the new config we have less failover destinations.
3) A lease-set comes (with the new config) for the lease token issued in (2).

This patch prevents this from happening by not using LeaseTokenMap if the number
of failover destinations changed.

An alternative approach could be to attach a lease-token to a config version.
As most reconfigs don't change the number of failover destinations, this would
cause more LeaseSet failures than necessary.

Reviewed By: jmswen

Differential Revision: D5906407

fbshipit-source-id: 2de2fa0e26b82f3dc021e4a20661abd4e2158893