Add support for custom failover route failover policy
Summary: For cross region failover, add the ability to specify a custom failover policy for FailoverRoute route handle. The policy is used when tao_xregion_failover json element is specified with value containing a bit offset for each child in the children element's array. The offset is used to set the bit in the region mask array to avoid requests from being stuck in a closed loop.

Reviewed By: alikhtarov

Differential Revision: D5381836

fbshipit-source-id: b3ec573338a89f5d2c88d66ffcd710e4b9003b2a