Do not copy pool name to DestinationRoute
Summary:
DestinationRoute is destroyed when ProxyConfig is destroyed.
We can use that premise to avoid copying the poolName to every DestinationRoute
we create. Instead, we can have a StringPiece to the poolName that is stored
in ProxyConfig itself.
This will save around ~52 bytes per connection (and speed up our config time).

Reviewed By: spalamarchuk

Differential Revision: D5883935

fbshipit-source-id: 1dda84336de18ba3ef6f3206a56d7f11722cf352