Smarter notifications, batching & scheduling, try 2
Summary:
Over the week I was thinking how to avoid double-scheduling `WriterLoop` (i.e. end of next loop). After jmswen fix to the messagequeue I realized that these things are related and we can use knowledge of them to improve each other, i.e. we ever need to set `State::EMPTY` if only we're going to block, similarly we don't need to flush every 2nd loop, but instead when we know that there's nothing more to do.

Try 1 caused prod issues due to a double queue drain, this diff fixes that. A couple points about the original issue:
The nature of a race is that mcrouter may attempt to drain the request queue twice (first time in runBeforeLoop callback, second in the notification event handler). Each of them will check if we need to flush something and are we going to perform blocking epoll_wait call, if that is true, we will schedule a flush callback. The issue arises when both decide to schedule it (they use the same callback object) and may override the flush list. Now the interesting question is when this going to happen.

1) For the event handler to fire, the request MUST arrive BEFORE we start draining in the runBeforeLoop callback (otherwise due to the flush callback, we would prevent the event from firing as an optimization).
2) For the event handler to schedule the flush callback, we must get some connection added into a flush list AND there should be no new requests (otherwise as an optimization, they would prevent us from scheduling the callback for at least one event loop).
3) The connection gets added to the flush list in three cases: request is sent, `connectSuccess` callback, goAway received.
4) Since we need this to happen between the first drain and the second drain, it needs to happen in libevent event, thus it cannot be a request sending path, because it's always happening on fibers and thus from a regular EventBase callback, goAway is not enabled, which leaves us with only `connectSuccess`.

Ultimately the problem is observed when we get a race between `connectSuccess` (rare event due to connections being closed very rarely) and queue event becoming active + other conditions met.

Reviewed By: jmswen

Differential Revision: D6001175

fbshipit-source-id: c99039b3ecf88e3611465f16f455d8236bea6eb1