Fix race destroying VirtualEventBase and EventBase
Summary:
This diff fixes a race that was happening when an exception was thrown when the server is starting to accept connections (e.g. port already in use).

Problem Description:
Mcrouter uses VirtualEventBase, and the actual EventBase is owned by AsyncMcServer.
When we spin-uw mcrotuer, it schedules a few things to run in the VirtualEventBase, which issues a new keep-alive token.
But when an exception is thrown by server when we are starting to accept connections, we haven't yet spawned EventBase threads.
So, when the exception propagates up, we try to destroy the VirtualEventBases from the main thread, and VirutalEventBase destructor blocks until all keep-alive tokens have been released.
As there's no thread looping the EventBase (to actually release the keep-alive tokens), mcrouter gets stuck.

Solution:
Destroy the EventBases from it's thread.

Reviewed By: andriigrynenko

Differential Revision: D6039264

fbshipit-source-id: 28da042dd09ec83ca017546b3d0e91d2acd11814