Improve shadowing of lease operations
Summary:
Currently, shadowing of lease-get/lease-set is broken in the following way:
* Send lease-get to primary and shadow.  Assume it's a miss on both, so both return valid lease tokens.
* Because there was a miss to the primary, the client will send a subsequent lease-set using the primary lease token that was issued.
* This lease-set will be propagated as-is to the shadow host, which is expecting a different token.  The lease-set may well fail; the original token also hangs around on the server side.  (The lease-set to the shadow may also be stale-stored, but that's another story.)

This diff fixes the above behavior by maintaining a map from primary token -> shadow token when a lease-get results in a miss to both primary and shadow.

Note that in theory, a single primary destination could have many shadow destinations, i.e., a primary token may be associated to more than one shadow token.  This limitation is not addressed in this diff.

Reviewed By: spalamarchuk

Differential Revision: D5984659

fbshipit-source-id: 01e34c7c361aad90d968e69d97fddfe8cffec683