Remove poolName from TkoLog event
Summary:
This patch fixes a race/ub in ProxyDestination.
We were updating poolName_ from config thread, while reading it from proxy thread. On top of that, we were using StringPiece in TkoLog, which could point to a destroyed string by time we actually tried to log.
This also saves around ~64 bytes of memory per connection (both openened and closed connections).

The poolName in ProxyDestination was just used for logging TKO events, and it was "best effort". ProxyDestination is shared among pools that point to the same host, so the poolName_ inside ProxyDestination would actually point to the name of last pool we saw when parsing configs.

Reviewed By: jmswen

Differential Revision: D5980071

fbshipit-source-id: 97897a42baa68c82acf1f75c51516636b8419d1e