Enable creating failover route with no failover destinations
Summary:
When creating failover route, if there's only one failover destination, we end up
short circuting the failover route. This prevents the failover policy from engaging resulting
in the cross region failover policy failing to set the region mask. This change makes it
such that the failover route is always created as long as there's one destination and zero
or more failover destinations.

Reviewed By: davidgoode

Differential Revision: D6997651

fbshipit-source-id: 59de92204dbfea57fe6d3347e7d26c468f16aeb7