Fix McStatsRequest sent through libmcrouter
Summary:
This fixes a bug when sending McStatsRequest sent to libmcrouter.
We were trying to create a std::string from nullptr (because commandargs stat
was empty as we don't have command line args in a library), which was throwing
an expcetion that wasn't being handled.
Now, if the value of the stat is nullptr and the type is string, we just return
empty string.
We are also catching exceptions and returning mc_res_local_error in case
something happens (instead of failing on an assert).

Reviewed By: glamtechie

Differential Revision: D7377253

fbshipit-source-id: fb84156cfe034ab1325f4bdbeb0f492084df6f29