Docker - j/Jobs ARG To Reduce Build Time
Context: With Docker 1.9 (Nov 2015), ARG can be used to provide
build-time arguments.

The mcrouter Dockerfile does not support the ability to provide
a build-time argument for the number of `make` jobs. Why this
matters:

| vCPU | -j arg passed | Time Elapsed |
| --- | --- | --- |
| 2 | none | 27m26.955s |
| 2 | -j2 | 18m20.238s |
| 4 | -j4 | 11m34.981s |

If `-j` is set to the number of vCPUs, the time it takes to build
mcrouter decreases significantly.

```
$ docker build -t mcrouter /home/amcrn/mcrouter/mcrouter/scripts/docker
```

would still default to `-j1`, whereas:

```
$ docker build --build-arg MAKE_JOBS=4 -t mcrouter \
    /home/amcrn/mcrouter/mcrouter/scripts/docker
```

would use `-j4`.

Closes: https://github.com/facebook/mcrouter/issues/97