Use Config struct for TestServer
Summary:
These constructors are a little easier to deal with if we define a
config struct with default parameters. This avoids having to define new factory
methods for each behavior we want (e.g. createWithCustomSslPaths) and makes it
easier to add new options and the ability to easily override them.

Reviewed By: andreazevedo

Differential Revision: D7405194

fbshipit-source-id: da50b3f25b4af62ef192ab36070ffa2b8125c095