Add an admin command (dp_setup_usecase) to precreate Zippy client
Summary:
This diff adds an admin command to precreate the zclient for a given
zippy use case. This allows us to tide over slow configerator accesses
during the first access and makes tests and setups less flakey/fragile.
This is similar in spirit to D7874920 which adds an admin interface for Zinc
to do the same.

Reviewed By: atejasvi

Differential Revision: D7892230

fbshipit-source-id: e838a6d50be4cea51a529467e0afd3816632a2e1