Carboncache integration test failing.
Summary:
The test that is failing is test_carbonlookaside_larger, this test does 2000
sets to mc, 2000 gets which it verifies gets the answer from the regular mc
server, then does 2000 gets again, which it verifies gets the replies back from
the carbonlookaside server. The test took > 11s to run and the TTL on the items
in this test was only 10s, hence they are expiring within 1 second as they are on
the TLRU.

No need for such a short TTL on these tests. Increasing it from 10s to 120s.I

Reviewed By: andreazevedo

Differential Revision: D7982629

fbshipit-source-id: a382adb524266ef3e47cf771fbba2d4241f60ab4