Use median for default server load
Summary:
This diff introduces a couple minor but nuanced changes to mcrouter
load balancer.

1) Previously the two-random choices algorithm had the possibility to
completely ignore the highest load server if the expiration timer did not
bump the load to a low enough default value. This has been addressed with
an unconditional expiration timer. After discussion with stuclar we have
made the weighted hashing algorithm use an uncoditional expiration timer as
well because we did not see a benefit in keeping it and it simlifies the
code so that someone doesn't come back a year later trying to figure out why
it exists.

2) The default load value to which a server is reset after hitting the
expiration timer has been changed to a dynamically tracked median. This
helps further address the situation in (1) and allows users of mcrouter
have a sane load "reset" value which might change throughout the day based
onchanging traffic patterns.

Reviewed By: andreazevedo

Differential Revision: D9374383

fbshipit-source-id: 4047913503c4400a2d99bdfcfd5d4f1eced27ad4