Support default shard split
Summary: Introduce the concept of a default shard split to ShardSplitter. Previously several places in ShardSplitter implicitly assumed the default shard split was 1. Now we return a pointer to defaultShardSplit_, which is a full ShardSplitInfo object. This allows graceful migration of the default shard split value from an old to new value just as per shard splits are also supported. The results of getShardSplit() still should be null checked, since the 2 parameter overload could still return nullptr, but in general getShardSplit() will now return a non-null value.

Reviewed By: andreazevedo

Differential Revision: D9112428

fbshipit-source-id: 2fcb6fd6c4a5a4ed62bc8246ee5b0f8d4560f1cf