Separate connect timeout from write timeout
Summary:
Allows users to separately specify connect timeout, and write/request timeouts.

The initial TCP timeout in linux is 1s, meaning that, if the initial SYN or
SYN-ACK (response) packets are lost/dropped, it will take at least 1 second for
TCP to retransmit the initial SYN packet.

With this change, it's possible to set the connect timeout to some value greater
than 1 second (because of the above), but still use sub-second timeout for
requests.

Reviewed By: glamtechie

Differential Revision: D9931746

fbshipit-source-id: 813fb207f15796d51ec07395607d0788ce383d1d