add ability to disable logging and customize reply result for error route
Summary:
add ability to customize reply result of ErrorRoute:
This diff should not affect the existed McRouter configs like:
   "ErrorRoute | custom msg"
By default, it returns the result of mc_res_local_error as before, and logs the reply.

This diff add the ability to specify the error reply result in the ErrorRoute config, for example:
   {
      "type": "ErrorRoute",

      /**
       * new optional field, specify the response result.
       * now only support <busy> and <local_error>. Default to <local_error>.
       */
      "reply_result": "busy",

      /**
       * new optional field, specify whether we should log the reply for the ErrorRoute. Default to true.
       */
      "enable_log": false,

      /**
       * optional error message in the reply. Default to empty string.
       */
      "response": "my message"
   }

ErrorRoute can then be used as an redirect route. The parent route can check the result from ErrorRoute, it's busy it can reroute to another route.

Plan to use it for StagingRoute.
When StagingRoute sees a busy reply from the route (ErrorRoute), it just stop sending next request to it.

Reviewed By: andreazevedo

Differential Revision: D10021288

fbshipit-source-id: ff4a77ca55a369c865caa2d327d153edf9c8ce2c