Increase no. of retries for WeightedCH3Hash
Summary:
SR uses WeightedCh3 hash for consistent hashing because is much faster
than most other consistent hashing algorithms out there.

We pass in a std::vector<double> of weights and a key to weightedCh3hash() method
which returns an index. This call can fail due to multiple reasons, which I have
updated the documentation with but SR had been relying on weight of the returned
index being 0 to consider it a failure which skews the probabilities.

This diff reduces the chances of same by increasing the no. of retries within
weightedCh3Hash method. Note that the outer loop of SR retries based on
svc_sticky_fallback_count doesn't work as for the case when weightedCh3Hash fails
returning an index with non-zero weight, we don't retry.

Reviewed By: andriigrynenko

Differential Revision: D10112971

fbshipit-source-id: 21aa2b120974e283af83b5b091d4724e5f7e0229