Allow for prefix matching for shard splits
Summary: Right now our shard splitting mechanism needs to find an exact match, this isn't the best if we want default splits per hostname prefix or shardmap. We should extend this so that if we fail to find an exact match, we fall back onto a longest prefix match based on the contents of the splits file passed to the MyRouterShardSplitRoute.

Reviewed By: jkedgar, andreazevedo

Differential Revision: D12844246

fbshipit-source-id: 527367628d4ef4840755feb90ff607f079caab9d