Set rlimit when using a custom number of connections
Summary:
When the user specifies a custom max number of connections, try to raise
rlimit so that we can actually reach that number of connections.
If we cannot update rlimit, drop the max number of connections down to what
we can actually have with the current rlimit setting, and return this new number
back to the user.

Reviewed By: stuclar

Differential Revision: D12854659

fbshipit-source-id: e41e3c33ea7f899f52ef1536cf1848658beb1695