Provide McServerSession& to the onConnectionAccepted callback
Summary:
Provides a reference to the McServerSession to the callback that is called
when a connection is accepted.
This diff also fixes a "bug" where, in some rare cases, we would call the
onConnectionAccepted callback, but never call the onConnectionClosed callback.
That would happen if there an exception was thrown when constructing the
McServerSession object.

Reviewed By: ngoyal

Differential Revision: D12916181

fbshipit-source-id: bafe62fc7d0f8cd502cd6f6e6f99d8a84fb009df