Implement copy avoidance on TX in AsyncMcServer
Summary:
Implement copy avoidance in AsyncMcServer using MSG_ZEROCOPY. This has been proven to be more efficient as it saves a packet copy on TX from userspace to kernel space. Its not however for free, copy avoidance is exchanging the cycles saved with the packet copy against the overhead of the page accounting and the notification back to a FD that needs to be read from userspace to indicate to the caller that they can free the buffer.  Kernel documentation states that this is only effective at writes > 10K due to the overhead of the notification management and the extra read that needs to be performed for each write (on the error queue).  https://www.kernel.org/doc/Documentation/networking/msg_zerocopy.rst In our case its a little more complex, given that we already use scatter-gather IO xmitting multiple IOvecs/pkt and also that optionally AsyncMcServer can xmit in batch mode.

TCP copy avoidance on TX has been implemented both for batch and single writes and is supported with and without SSL. It permits a threshold to be configured, which if an IOvec in a batch exceeds this threshold, the system call will be made with the zero copy option enabled. If it the threshold is not exceed, xmit will be performed normally with the normal packet copy.  Its very much up to the discretion of the individual usecase to determine if this will be a performance boost in production. Benchmarks below show a performance improvement however memcache in particular did not (CCB is the only pool using AsyncMcServer and large objects, which not only has very low CPU anyway but not enough significantly large objects to make this worthwhile).  Other AsyncMcServer usecases, such as DSI may benefit.

Note: In order to use tcp zero copy, it may be necessary to increase the ulimit for max locked memory, given that page pinning is being used. This currently defaults to 16384K.

ulimit -l unlimited

It is also recommended to use the maxReadsPerEvent option in AsyncMcServer worker to ensure that the error queue is not starved due to a high number of reads on any given socket.

Reviewed By: andreazevedo

Differential Revision: D10287749

fbshipit-source-id: 526134b685daf6f1e9875c03f7ee6237a07b22c7