Fix flaky tests
Summary:
Fixed flaky tests that were failing sporadically, partilarly on machines under
some kind of pressure.
There are still lots of room for flakiness (especially in regard to tests that
rely on time). I improved some of these, but they are not 100% reliable - we
need to completely stop relying on timing for that.

Reviewed By: stuclar

Differential Revision: D13018043

fbshipit-source-id: c02be780ee2070124bca7a947a443b3e6fa4b635