WeightedCh4HashFunc, using weighted furc hash
Summary:
The stacked diff D13018455 introduced a new weighted consistent hash function, furcWeightedHash().
In this diff, the new hash function is exposed as a mcrouter has function called WeightedCh4HashFunc.
I'm not sure about the name - it's not based on CH3 (and there is no CH4 AFAIK), but it needed to be named with some association to WeightedCh3HashFunc because it's partially compatible (if weights are all 1.0).

Reviewed By: glamtechie

Differential Revision: D13030212

fbshipit-source-id: f6c338b416143e6fdd7ba99256539334c5aa292b