weighted furc hash
Summary:
The main consistent hashing function we have today is WeightedCh3HashFunc.
WeightedCh3 works by invoking furc_hash on the key, which selects a bucket index, and then it runs SpookyHash to enforce the weights. In case the key falls in the rejected segment (above the weight threshold), WeightedCh3 tries different variations of the key, sending them to furc_hash and SpookyHash in a loop, until acceptance (or max attempts).
This algorithm can be made more efficient, if we could have the furc consistent hash function enforce the weights.
The new algorithm, weightedFurcHash, takes the weight array as input and enforces the weight against the next 32 bits generated in the murmur hash stream. In case of rejection, it continues to additional furc hash index iterations, continuing on the same murmur hash stream, using the hash cache as a cyclic buffer.
There are two main inefficiencies that are solved. With weightedFurcHash, there is

* No need to run two different hashing functions on the same key
* No need to loop with different key variations

NOTE: The resulting routing function will only be compatible with WeightedCh3 in cases where the weights are all 1, so any change in production will have to go through gradual migration process.

Reviewed By: andriigrynenko

Differential Revision: D13018455

fbshipit-source-id: 9f5110c5561195d8c9bb24ec1ebd290e7387053b