MCRouter should error when a weight is not in range [0,1]
Summary:
When a `weight` of a `HashRoute` is not in range `[0, 1.0]`, `mcrouter` will reject the config with an error:  `WeightedChHashFunc: weight must be in range [0, 1.0]`. It will then continue running with the old config prior to the change.

The current behavior is an assertion failure, which is compiled out in production builds - making these bad configuration changes hard to detect. This change ensures it is caught at configuration time.

Reviewed By: orishu

Differential Revision: D13503667

fbshipit-source-id: 3587bdf5f92d3b15b3f6df2a3a1ccf7671e266e8