Fix broken Mcrouter tests as a result of expecting wrong exptime
Summary:
When shipping D13629181, for some reason `mcrouter` tests did not run. This led to a bunch of silent failures having to do with the new default exptime of 1 week.

In this diff, because we aren't testing that there's a default exptime, we instead supply an exptime of `100` for everything without an exptime. This will avoid us having to special case OSS memcache (which does not have a default exptime of 1 week).

Reviewed By: andreazevedo

Differential Revision: D13893998

fbshipit-source-id: 4c3143c701b37934a156057dee65af8b561d2735