Add Memcache thrift interface
Summary: As per title adds a Thrift interface for Memcache. Note, we need to figure out how to actually name the methods, because we cannot use `set` or `delete`, because first one is reserved in Thrift, second one is reserved in Thrift and C++. So I'm open for suggestions, we can do something like `setKV`, `deleteKey`, but we'd need to name others in some similar fashion.

Reviewed By: andreazevedo

Differential Revision: D14021928

fbshipit-source-id: 64a276771108b068aaa084ceff8ea9306394ba1e