Build zstd before folly and use a release version instead of dev
Summary:
Currently we build folly without zstd and then build zstd for fbthrift. This results in a side effect that upon rebuild, folly will pick up the zstd and trigger rebuild.

Interestingly, this hits another problem - we're using a dev version of zstd that may contain breaking API changes. I'm pinning 1.3.7 as 1.3.8 actually introduced an API change that is not handled in folly yet.

Reviewed By: andreazevedo

Differential Revision: D14205367

fbshipit-source-id: 495eba7666d42e6c4076b8edf260d43ee4e912a3