Fix StagingRoute bug on all set operations where empty values are set to the staging side.
Summary:
The staging route will incorrectly set empty values to the staging side after setting the warm side.  This affected `set`, `lease-set`, and `cas` operations.  The problem is that the route was incorrectly using the reply from the warm route to construct the request packet for the staging route.  The reply from memcached will never include the actual value in the response on a write operation.

This change will fix it such that it will construct a set request object based on the original request (set/lease-set/cas) as well as update the unit tests to also check the values seen.

Reviewed By: andreazevedo

Differential Revision: D14320266

fbshipit-source-id: c1effe99f121159e9d99e0d727d19a6dab3ae3fa