Improve support for custom list of children
Summary:
Improve the support for custom list of children in route handles.
Now, route handles have to explicitly pop the children list when it's no longer
needed. This allows for the config to re-use the same list of children twice
in the same context (see attached unit test for more details).
This also changes the children's list string from "%children_list%" to
"$children_list$". With this, it's no longer necessary to escape the '%' (i.e.
before this change it was necessary to use "\\%children_list\\%" to access the
children's list. Now, just "$children_list$" will achieve the same thing).

Reviewed By: arbass22

Differential Revision: D14884718

fbshipit-source-id: 77587c04a1ef93e9d197fc95030b48f4ec3ce4a3