Add first version of thrift transport
Summary:
Add support to the very first version of thrift transport. At the moment, thrift
transport just returns the default reply for a given request (like NoOpTransport
does).
If ThriftTransport::sendSync() is called on IDLs that don't have thrift enabled,
a logic exception is thrown with a message saying that the given IDL doesn't
support thrift.

Reviewed By: glamtechie

Differential Revision: D14846328

fbshipit-source-id: 54fc02b67eab443950a3e7b89d04337a5c7e835d