Thread affinity implementation for libmcrouter to reduce number of connections between client/server
Summary: CarbonRouterClient now supports a thread_affinity mode where the CarbonRouterClient will run a traverse on the request to determine the proxy thread that should be used to establish a connection to the server. This will decrease the number of connections between client and server by O(num_proxies) when using the InternalCarbonConnectionImpl. Note that this optimization applies to libmcrouter only.

Reviewed By: andreazevedo

Differential Revision: D14994321

fbshipit-source-id: 4ae9b902bdb7a5fc0209c9f18e5cddc918f6494f