Store hash in carbon::Key to enable reuse in proxy threads
Summary:
Store hash in carbon::Key to enable reuse in proxy threads. This is an optimization when using thread affinity that , where possible, offloads the proxy thread from having to recalculate the hash.

The hash result is reused when there is no salt, the pool is of the same size and the hash function is the same.

Reviewed By: andreazevedo

Differential Revision: D14994315

fbshipit-source-id: 4276e930f67df57c3d92138e27df9f10be680650