Remove usages of folly::ssl::setLockTypes
Summary:
This was always a dangerous API to use, and the performance issues that this
API was trying to address have been addressed in OpenSSL 1.1, which is now
the default in platform007.

The last time we tried to remove setLockTypes, this caused S170138. However,
since then, login service has been on platform007 for quite some time, so this
call has been a no-op since.

Reviewed By: knekritz

Differential Revision: D15324668

fbshipit-source-id: acb41579f5d079cee4babe5e69353afa77c94da7