Optimize CarbonRouterClient construction.
Summary:
A couple of optimization with regards to the construction of CarbonRouterClient.

The first one is the fact that, on each construction of CarbonRouterClient we
were locking a weak_ptr twice. We were basically executing the following cycle:
weak_ptr -> shared_ptr -> weak_ptr -> shared_ptr.
Now, we just lock it once (weak_ptr -> shared_ptr -> weak_ptr).

The second optimization is that we were copying the vector of pointers to Proxy
from the router to the client. Now we just store a const& at the client.

Reviewed By: glamtechie

Differential Revision: D15463473

fbshipit-source-id: 4bbbe77bb3e491ecef3c9ae4a8ded74a4bd0b529