Improve batching of multi-request send()
Summary:
This diff delays notifying the event base thread for when sending multiple
requests at once using CarbonRouterClient.
Before this change, the event base thread was notified once for every request.
Now, with this change, the event base thread is only notified once at the end
of each batch.
The end of the batch is usually right afeter we finished enqueueing the last
request passed to CarbonRouterClient::send(). But that's not always the case.
If, for example, we would block (because of the semaphore), we end the batch
right before blocking.

Reviewed By: stuclar

Differential Revision: D15466524

fbshipit-source-id: a4d68d86179381bf5518f1eff6995af64207f516