Optimize CarbonRouterClient when thread-affinity is enabled
Summary:
This change optimizes CarbonRouterClient when ThreadMode is
AffinitizedRemoteThread.
Before this change, in order to pick the best Proxy to pass the request to, we
were using the foreachPossibleClient API. Now, we are now traversing the routing
tree directly from CarbonRouterClient.
With this, we avoid allocating a ProxyRequestContext and a McrouterFiberContext.
In terms of perf improvements this saves, for every request, a couple of heap
allocations/destructions, and one atomic operation.

Reviewed By: stuclar

Differential Revision: D15503135

fbshipit-source-id: 27c2f99c1450d777d56c8800e9028f906374e8ff