Add ThriftTransport constructor taking plain EventBase
Summary:
Currently, the generated `ThriftTransport` constructor accepts a `VirtualEventBase&`. This is for consistency
with `AsyncMcClient` code. However, currently `ThriftTransport` doesn't use the virtual-ness of `VirtualEventBase` at all
and simply directly uses the backing `EventBase`. For tests, it would be convenient to be able to construct a `ThriftTransport`
directly from `EventBase`, which is what this diff adds.

Reviewed By: stuclar

Differential Revision: D17056270

fbshipit-source-id: f968417e95952f48c58c087bf0d3d4035ee46267