make enable/disable shadow shards independent from shards for EagerShardSelectionShadowRoute
Summary:
There is a corner case where EagerShardSelectionShadowRoute does not route shadow shard requests when shards is empty.

Before this fix, we optimized SelectionRoute to ErrorRoute when shards is empty.
After this fix, we optimized SelectionRoute to ErrorRoute only when both shards and shardow_shards are empty.

(Note: this ignores all push blocking failures!)

Reviewed By: stuclar

Differential Revision: D17741973

fbshipit-source-id: 5c1e49e113acdcc8ea25d248ca6ba6162cacdd32