Client authorization
Summary:
McRouter client now has the ability to authorize the TLS service identity of the servers that it connects to with the expected identity configured in ssl_service_identity.

The following flavor options have been added:

    ssl_service_identity_authorization_log: "The configured service identity of the client is compared against the service identity of the server in the peer certificate. Log if they match."

    ssl_service_identity_authorization_enforce: "The configured service identity of the client is compared against the service identity of the server in the peer certificate. Fail to connect if they do not match."

The stat "num_authorization_failures" has also been added which tracks the number of authentication failures.

Reviewed By: jmswen

Differential Revision: D17993322

fbshipit-source-id: 69fac1f7f62d1d87a3a9e5f5f4c1c6826863cf52