Fix bug in AsyncTlsToPlaintextSocket
Summary:
This diff fixes an obvious bug where, currently, an always-null `readCallback_` will be dereferenced in case `close()`/`closeNow()` is somehow called before `connectSuccess()`/`connectErr()` fires. This is mostly out of caution; we haven't observed this behavior anywhere.

This diff also 1) adds an `UNLIKELY` hint to the somewhat hot `writeChain` function and 2) makes `getReadCallback()` more usable, working even if the socket is not in the connected state. (This function is mostly for testing AFAIK so this is a minor but straightforward improvement.)

Reviewed By: stuclar

Differential Revision: D18598726

fbshipit-source-id: bfce1967db85b9428afd5c793d81cdfba8126c40