Reduce memory usage in ProxyDestinationMap
Summary:
With heterogeneous find available in `folly::F14` containers, we can store pointer to ProxyDestination and use lookup with a custom key. There is no need to store string key anymore.
Furthermore, there is no need to store weak pointers. Since for the entire lifetime of ProxyDestination we store it in a map and in destructor we remove it from the map, storing raw pointer should be fine. This also removes races with ProxyDestination destruction on stats thread.

Reviewed By: stuclar

Differential Revision: D18297873

fbshipit-source-id: 026e4f202594f8f3007605e2d8be0a4e46b5ea8b