Allow overriding toggle option configured in flavor
Summary:
Currently with the way toggle options (eg. `--acl-checker-enable`) work in mcrouter standalone,
passing the command line parameter can only set the option, but not reset it if it was set in the
flavor file.

This diff allows toggle options to have an optional command-line parameter to override the value
set in a flavor config (eg. `--acl-checker-enable=0`).

Reviewed By: stuclar

Differential Revision: D18902759

fbshipit-source-id: 61251c004161ac8934845c55539bd2ca2a335805