make ProxyDestinationKey.cpp more portable
Summary:
On some systems where we build HHVM packages, `std::hash` is not available for enums (specifically, GCC 5.5 doesn't support it despite claiming to support C++17) so we get compilation errors here.

These `static_cast`s (backported from how the hash is declared for enums in newer GCC's header files) would fix that.

If you're opposed, that's fine, I can abandon this and keep it as a HHVM-specific patch in our build system; but maintaining these patches is a non-trivial burden so anything like this that we can fix more generally helps us quite a bit.

Reviewed By: pavlo-fb

Differential Revision: D19279530

fbshipit-source-id: 4861c4dbd22768680ae3aafd6ba5b53ca310fc7c